added connectivity, some changes

+++ lib/data/repo/auth_repo.dart
@@ -1,19 +1,41 @@
+import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter_test_app/data/models/user.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'dart:async';
+import 'package:connectivity/connectivity.dart';
 
 class AuthRepo {
+  bool network = false;
+  String messageNetworkError =
+      "Something went wrong. Please check your internet and try again";
+
+  Future<void> checkInternet() async {
+    print("check int");
+    var result = await Connectivity().checkConnectivity();
+    if (result == ConnectivityResult.none) {
+      print("not int $ConnectivityResult");
+      network = false;
+    } else if (result == ConnectivityResult.mobile || result == ConnectivityResult.wifi) {
+      print("have int $ConnectivityResult");
+      network = true;
+    } else {
+      network = false;
+      print("error $ConnectivityResult");
+    }
+  }
+
   Future<List<ListUsers>> getUsers() async {
     try {
-      //    bool network = false;
-//    if (!network) {
-//      throw PlatformException(
-//        code: "Test error code",
-//        message: "Test error message",
-//      );
-//    }
+      checkInternet();
+      if (!network) {
+        throw PlatformException(
+          code: "Test error code",
+          message: messageNetworkError,
+        );
+      }
 
       QuerySnapshot querySnapshot =
           await Firestore.instance.collection("users").getDocuments();
@@ -33,13 +55,13 @@ class AuthRepo {
     String email,
     String password,
   ) async {
-    //    bool network = false;
-//    if (!network) {
-//      throw PlatformException(
-//        code: "Test error code",
-//        message: "Test error message",
-//      );
-//    }
+    checkInternet();
+    if (!network) {
+      throw PlatformException(
+        code: "Test error code",
+        message: messageNetworkError,
+      );
+    }
 
     try {
       final AuthResult authResult =
@@ -62,9 +84,21 @@ class AuthRepo {
     int age,
   ) async {
     try {
-      Firestore.instance
-          .collection("users").document()
-          .setData({
+//      var result = await Connectivity().checkConnectivity();
+//      if (result == ConnectivityResult.none) {
+//        print("throw PlatformException");
+//        throw PlatformException(
+//          code: "Test error code",
+//          message: messageNetworkError,
+//        );
+//      }
+
+      throw PlatformException(
+        code: "Test error code",
+        message: messageNetworkError,
+      );
+
+      Firestore.instance.collection("users").document().setData({
         'firstName': firstName,
         'lastName': lastName,
         'age': age,
@@ -78,13 +112,13 @@ class AuthRepo {
     String email,
     String password,
   ) async {
-    //    bool network = false;
-//    if (!network) {
-//      throw PlatformException(
-//        code: "Test error code",
-//        message: "Test error message",
-//      );
-//    }
+    checkInternet();
+    if (!network) {
+      throw PlatformException(
+        code: "Test error code",
+        message: messageNetworkError,
+      );
+    }
 
     try {
       final AuthResult authResult =
@@ -104,6 +138,14 @@ class AuthRepo {
   Future<void> deleteData(
     DocumentSnapshot document,
   ) async {
+    checkInternet();
+    if (!network) {
+      throw PlatformException(
+        code: "Test error code",
+        message: messageNetworkError,
+      );
+    }
+
     try {
       await document.reference.delete();
     } catch (error) {
@@ -115,6 +157,14 @@ class AuthRepo {
     DocumentSnapshot document,
     String newName,
   ) async {
+    checkInternet();
+    if (!network) {
+      throw PlatformException(
+        code: "Test error code",
+        message: messageNetworkError,
+      );
+    }
+
     try {
       await document.reference.updateData({'names': newName});
     } catch (error) {
